push for remote access

+++ ae_on_heads_w_keith/z_sae.py
@@ -158,7 +158,7 @@ class AutoEncoder(nn.Module):
         else:
             self.cached_acts = None
         if record_activation_frequency:
-            self.activation_frequency += (acts > 0).mean(dim=0)
+            self.activation_frequency += (acts > 0).float().mean(dim=0)
             self.steps_since_activation_frequency_reset += 1
         return x_reconstruct
